Initial drainflow file
Added file with initial drain flows set to zero. This file is then copied and renamed to the drain file used by the SWBM to determine inflow to the Eastside slough from the tile drains in the discharge zone.  Without this file the initial values in the drain file would be from a previous run, which can cause errors if parameter values are changed.